02-test/compile-and-test-one-board: add script to test board
Add a script to compile and execute test on a board and produce file results.

This script handles building all applications and tests for one board and also
execute tests if they are available.

An incremental build can selected using `--incremental` to not rerun successful
compilation and tests. But then it should be run on a fixed version of the
repository as no verification is done if results are up to date with the RIOT
repository.

It by defaults finds all tests in `examples` and `tests` but list of tests can
be provided by command line options with also an exclude list, to for example
not rerun a long failing test every time.

It is a temporary solution until an equivalent is implemented in the build
system. It is also a showcase of what could be dummy file output for
compilation and tests.

Example
-------

By default it should be run as

    ./compile_and_test_for_board.py path_to_riot_directory board_name [results]