02-tests:compile_and_test_for_board: allow setting 'all' and 'test'
Allow overwriting compile 'clean all' and 'test' commands.

This will allow running for example 'scan-build-analyze' for building
and also other test commands like 'robottest', 'pytest',
'archive-check' as currently exist in some PRs.

It still currently hardwritten how to finding if there is a test
as it is not integrated in RIOT or if a board is supported for the
application.