release-tests: mark IoT-LAB runs as flaky
Oftentimes tests fail only because there is currently a lot of noise in
the IoT-LAB testbed due to ongoing experimentation. This change allows
for those tests to be repeated on failure after 30 seconds up to 3
times.

Experimental tests marked with XFAIL are skipped for this as they will
be marked as skipped on failure anyway and that should not be hidden.

Signed-off-by: Martine Lenders <m.lenders@fu-berlin.de>